% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioage_calc.R
\name{bioage_calc}
\alias{bioage_calc}
\title{bioage_calc}
\usage{
bioage_calc(data, age, biomarkers, fit = NULL, s_ba2 = NULL)
}
\arguments{
\item{data}{The dataset for calculating KDM bioage}

\item{age}{A character vector (length=1) indicating the name of the variable for chronological age}

\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating KDM bioage}

\item{fit}{An S3 object for model fit. If the value is NULL, then the parameters to use for training KDM bioage are calculated}

\item{s_ba2}{A particular fit parameter. Advanced users can modify this parameter to control the variance of bioage}
}
\value{
An object of class "bioage". This object is a list with two elements (data and fit)
}
\description{
Calculate Klemera-Doubal Method (KDM) biological age
}
\details{
Calculate KDM biological age
}
\examples{
#Train KDM bioage parameters
train = bioage_calc(nhanes3,age="age",
                    biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"))

#Use training data to calculate KDM bioage
bioage = bioage_calc(nhanes,age="age",
                     biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"),
                     fit=train$fit,
                     s_ba2=train$fit$s_ba2)

#Extract bioage dataset
data = bioage$data


}
