% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioage_nhanes.R
\name{bioage_nhanes}
\alias{bioage_nhanes}
\title{bioage_nhanes}
\usage{
bioage_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating bioage}
}
\value{
An object of class "bioage". This object is a list with two elements (data and fit)
}
\description{
Calculate Klemera-Doubal Method (KDM) Biological Age Using NHANES Dataset
}
\details{
Calculate Klemera-Doubal Method Biological Age Using NHANES Dataset
}
\examples{
#Calculate KDM bioage
bioage = bioage_nhanes(biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"))

#Extract bioage dataset
data = bioage$data


}
