% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdmfuns.R
\name{extract_data}
\alias{extract_data}
\title{Extracts dataframe with age, controls, biomarkers, and biological age calculation from trianed data (using kdm_calc function).}
\usage{
extract_data(kdmobj)
}
\arguments{
\item{kdmobj}{A kdm object estimated using the function kdm_calc.}
}
\value{
A dataframe with parameters and biological age.
}
\description{
Extracts dataframe with age, controls, biomarkers, and biological age calculation from trianed data (using kdm_calc function).
}
\examples{
#(not run)
#Train biological age parameters
train = kdm_calc(nhanes,agevar='age',
  biomarkers=c('sysbp','totchol','bun','cmv','mcv'))

newdata = extract_data(train)
}
