% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdmfuns.R
\name{extract_fit}
\alias{extract_fit}
\title{Extracts and summarizes estimates for fitting biological age from a kdm object.}
\usage{
extract_fit(kdmobj)
}
\arguments{
\item{kdmobj}{A kdm object estimated using the function kdm_calc.}
}
\value{
A dataframe with trained parameters.
}
\description{
Extracts and summarizes estimates for fitting biological age from a kdm object.
}
\examples{
#(not run)
#Train biological age parameters
train = kdm_calc(nhanes,agevar='age',
  biomarkers=c('sysbp','totchol','bun','cmv','mcv'))

myfit = extract_fit(train)
}
