% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_calc.R
\name{hd_calc}
\alias{hd_calc}
\title{hd_calc}
\usage{
hd_calc(data, reference, biomarkers)
}
\arguments{
\item{data}{A projection dataset.}

\item{reference}{A training dataset.}

\item{biomarkers}{A character vector indicating the names of the biomarkers included in the HD algorithm.}
}
\value{
An object of class "hd". This object is a list with two elements (data and fit). The dataset can be drawn by typing 'data'. The model can be drawn by typing 'fit'.
}
\description{
Project HD algorithm onto new data.
}
\details{
For HD algorithm., the constructed variable is based on a malhanobis distance statistic, which is theoretically the distance between observations and a hypothetically healthy, young cohort. You need to train separately for men and women who are between the ages of 20 and 30 and not pregnant, and have observe biomarker data within clinically accpetable distributions.
}
\examples{
#HD using NHANES
hd = hd_calc(NHANES4, NHANES3,
             biomarkers=c("albumin_gL","lymph","mcv","glucose_mmol",
             "rdw","creat_umol","lncrp","alp","wbc"))

#Extract HD dataset
data = hd$data


}
