% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_nhanes.R
\name{hd_nhanes}
\alias{hd_nhanes}
\title{hd_nhanes}
\usage{
hd_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating HD}
}
\value{
An object of class "hd". This object is a list with two elements (data and fit)
}
\description{
Calculate Homeostatic Dysregulation (HD) Using NHANES Dataset
}
\details{
Calculate Homeostatic Dysregulation Using NHANES Dataset
}
\examples{
#Calculate HD
hd = hd_nhanes(biomarkers=c("albumin","lymph","mcv","glucose","rdw","creat","lncrp","alp","wbc"))

#Extract HD dataset
data = hd$data


}
