% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdm_nhanes.R
\name{kdm_nhanes}
\alias{kdm_nhanes}
\title{kdm_nhanes}
\usage{
kdm_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the biomarkers included in the KDM Biological Age algorithm.}
}
\value{
An object of class "kdm". This object is a list with two elements (data and fit). The dataset can be drawn by typing 'data'. The model can be drawn by typing 'fit'.
}
\description{
Train KDM algorithm in NHANES III and project into NHANES IV.
}
\details{
Training KDM Biological Age algorithm using the NHANES III and projecting into NHANES IV dataset.
}
\examples{
#KDM using NHANES
kdm = kdm_nhanes(biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"))

#Extract KDM dataset
data = kdm$data


}
