% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoage_calc.R
\name{phenoage_calc}
\alias{phenoage_calc}
\title{phenoage_calc}
\usage{
phenoage_calc(data, biomarkers, fit = NULL, orig = FALSE)
}
\arguments{
\item{data}{A projection dataset.}

\item{biomarkers}{A character vector indicating the names of the biomarkers included in the Phenotypic Age algorithm.}

\item{fit}{An S3 object for model fit. If the value is NULL, then the parameters to use for training Phenotypic Age are calculated.}

\item{orig}{TRUE to compute the origianl Phenotypic Age.}
}
\value{
An object of class "phenoage". This object is a list with two elements (data and fit). The dataset can be drawn by typing 'data'. The model can be drawn by typing 'fit'.
}
\description{
Project Phenotypic Age algorithm onto new data.
}
\details{
Projecting Phenotypic Age algorithm onto new data.
}
\examples{
#Train using the NHANES III
train = phenoage_calc(NHANES3,
                      biomarkers = c("albumin_gL","lymph","mcv","glucose_mmol",
                      "rdw","creat_umol","lncrp","alp","wbc"))

#Project into the NHANES IV
phenoage = phenoage_calc(NHANES4,
                         biomarkers = c("albumin_gL","lymph","mcv","glucose_mmol",
                         "rdw","creat_umol","lncrp","alp","wbc"),
                         fit = train$fit)

#Extract phenoage dataset
data = phenoage$data


}
