% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoage_nhanes.R
\name{phenoage_nhanes}
\alias{phenoage_nhanes}
\title{phenoage_nhanes}
\usage{
phenoage_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating phenoage}
}
\value{
An object of class "phenoage". This object is a list with two elements (data and fit)
}
\description{
Calculate Levine's Phenotypic Age using NHANES dataset
}
\details{
Calculate Levine's Phenotypic Age using NHANES dataset
}
\examples{
#Calculate phenoage
phenoage = phenoage_nhanes(biomarkers=c("albumin_gL","lymph","mcv","glucose_mmol",
                           "rdw","creat_umol","lncrp","alp","wbc"))

#Extract phenoage dataset
data = phenoage$data


}
