% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ba.R
\name{plot_ba}
\alias{plot_ba}
\title{plot_ba}
\usage{
plot_ba(data, variables)
}
\arguments{
\item{data}{The dataset for plotting correlations}

\item{variables}{A character vector indicating the names of the biological aging measures}
}
\description{
Plot Correlations Between Biological Aging Measures and Chronological Age
}
\details{
Plot Correlations Between Biological Aging Measures and Chronological Age
}
\examples{
#Calculate phenoage
f1 = plot_ba(data = data, variables = c("bioage", "phenoage", "hd"))

f1

}
