% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdmfuns.R
\name{project}
\alias{project}
\title{Project from a training dataset into a validation dataset using kdm or hd algorithms.}
\usage{
project(training_data, projection_data, method = "hd", biomarkers, ...)
}
\arguments{
\item{training_data}{The dataset for calculating biological age or hd.}

\item{projection_data}{The dataset for projecting biological age or hd into.}

\item{method}{A character variable of either 'hd' or 'kdm' to identify the}

\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating biological age.}

\item{filter}{a list with biomarker names that identifies any restrictions in training data. See vignette or data description for example of use.}
}
\value{
An object of class 'kdm' or 'hd'. This object is a list with two elements (data and fit),
and two methods (extract_data and extract_fit). Data is limited to the "projection" dataset. Fit is calculate from "training" dataset.
}
\description{
Project from a training dataset into a validation dataset using kdm or hd algorithms.
}
\examples{
#(not run)
#Train biological age parameters
ndata = project(training_data=nhanes,projection_data=holdout,
                agevar='age',
                biomarkers=c('sysbp','totchol','bun','cmv','mcv'))

}
