% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_health.R
\name{table_health}
\alias{table_health}
\title{table_health}
\usage{
table_health(data, agevar, outcome)
}
\arguments{
\item{data}{The dataset for plotting corplot}

\item{agevar}{A character vector indicating the names of the interested biological age measures}

\item{outcome}{A character vector indicating the name of the interested current health status outcomes}
}
\description{
Association with current health status outcomes
}
\details{
Association with current health status outcomes, adjusting for chronological age and gender and stratified by age
}
\note{
Chronological age and gender variables need to be named "age" and "gender"
}
\examples{
table2 = table_health(nhanes,
                      agevar = c("bioage_advance0","phenoage_advance0",
                               "bioage_advance","phenoage_advance",
                               "hd","hd_log"),
                      outcome = c("health","adl","lnwalk","grip_scaled"))

table2

}
