% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ses.R
\name{table_ses}
\alias{table_ses}
\title{table_ses}
\usage{
table_ses(data, agevar, exposure, label)
}
\arguments{
\item{data}{A dataset with projected biological aging measures for analysis.}

\item{agevar}{A character vector indicating the names of the biological aging measures.}

\item{exposure}{A character vector indicating the name of the socioeconomic circumstances.}

\item{label}{A character vector indicating the labels of the biological aging measures.}
}
\value{
The result is a list with two elements (table and  n). The regression table can be drawn by typing 'table'. The sample size table can be drawn by typing 'n'.
}
\description{
Associations of socioeconomic circumstances measures with measures of biological aging.
}
\details{
Coefficients are from linear regressions of biological aging measures on measures of socioeconomic circumstances. KDM Biological Age and Levine Phenotypic Age measures were differenced from chronological age for analysis (i.e. values = BA-CA). These differenced values were then standardized to have M=0, SD=1 separately for men and women within the analysis sample. Socioeconomic circumstances measures were standardized to M=0, SD=1 for analysis so that effect-sizes are denominated in terms of a 1 SD unit improvement in socioeconomic circumstances.
}
\note{
Chronological age, gender, and race/ethnicity variables need to be named "age", "gender", and "race".
}
\examples{
table3 = table_ses(data,
                   agevar = c("kdm_advance0","phenoage_advance0",
                               "kdm_advance","phenoage_advance",
                               "hd","hd_log"),
                   exposure = c("edu","annual_income","poverty_ratio"),
                   label = c("KDM\nBiological Age\nAdvancement",
                             "Levine\nPhenotypic Age\nAdvancement",
                             "Modified-KDM\nBiological Age\nAdvancement",
                             "Modified-Levine\nPhenotypic Age\nAdvancement",
                             "Homeostatic\nDysregulation",
                             "Log\nHomeostatic\nDysregulation"))

table3$table
table3$n

}
