% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_calc.R
\name{hd_calc}
\alias{hd_calc}
\title{hd_calc}
\usage{
hd_calc(data, reference, biomarkers)
}
\arguments{
\item{data}{The dataset for calculating HD}

\item{reference}{The reference dataset for calculating HD}

\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating HD}
}
\value{
An object of class "hd". This object is a list with two elements (data and fit)
}
\description{
Calculate Homeostatic Dysregulation (HD)
}
\details{
Calculate Homeostatic Dysregulation
}
\examples{
#Calculate HD
hd = hd_calc(nhanes,nhanes3,
             biomarkers=c("albumin_gL","lymph","mcv","glucose_mmol","rdw","creat_umol","lncrp","alp","wbc"))

#Extract bioage dataset
data = hd$data


}
