% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_nhanes.R
\name{hd_nhanes}
\alias{hd_nhanes}
\title{hd_nhanes}
\usage{
hd_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the biomarkers included in the HD algorithm.}
}
\value{
An object of class "hd". This object is a list with two elements (data and fit). The dataset can be drawn by typing 'data'. The model can be drawn by typing 'fit'.
}
\description{
Train HD algorithm in NHANES III and project into NHANES IV.
}
\details{
Training HD algorithm using the NHANES III (1988 - 1994) and projecting into NHANES IV (1999 - 2018) dataset. For this function, NHANES III included men and women who are between the ages of 20 and 30, and have observe biomarker data within clinically acceptable distributions.
}
\examples{
#HD using NHANES
hd = hd_nhanes(biomarkers=c("albumin","lymph","mcv","glucose","rdw","creat","lncrp","alp","wbc"))

#Extract HD dataset
data = hd$data


}
