% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdm_calc.R
\name{kdm_calc}
\alias{kdm_calc}
\title{kdm_calc}
\usage{
kdm_calc(data, biomarkers, fit = NULL, s_ba2 = NULL)
}
\arguments{
\item{data}{The dataset for calculating KDM Biological Age}

\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating KDM Biological Age}

\item{fit}{An S3 object for model fit. If the value is NULL, then the parameters to use for training KDM Biological Age are calculated}

\item{s_ba2}{A particular fit parameter. Advanced users can modify this parameter to control the variance of kdm}
}
\value{
An object of class "kdm". This object is a list with two elements (data and fit)
}
\description{
Calculate Klemera-Doubal Method (KDM) Biological Age
}
\details{
Calculate KDM Biological Age
}
\examples{
#Train KDM kdm parameters
train = kdm_calc(NHANES3,
                    biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"))

#Use training data to calculate KDM Biological Age
kdm = kdm_calc(NHANES4,
                     biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"),
                     fit=train$fit,
                     s_ba2=train$fit$s_ba2)

#Extract kdm dataset
data = kdm$data


}
