% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdm_nhanes.R
\name{kdm_nhanes}
\alias{kdm_nhanes}
\title{kdm_nhanes}
\usage{
kdm_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the variables for the biomarkers to use in calculating KDM Biological Age}
}
\value{
An object of class "kdm". This object is a list with two elements (data and fit)
}
\description{
Calculate Klemera-Doubal Method (KDM) Biological Age using NHANES dataset
}
\details{
Calculate KDM Biological Age using NHANES dataset
}
\examples{
#Calculate KDM Biological Age
kdm = kdm_nhanes(biomarkers=c("fev","sbp","totchol","hba1c","albumin","creat","lncrp","alp","bun"))

#Extract kdm dataset
data = kdm$data


}
