% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoage_nhanes.R
\name{phenoage_nhanes}
\alias{phenoage_nhanes}
\title{phenoage_nhanes}
\usage{
phenoage_nhanes(biomarkers)
}
\arguments{
\item{biomarkers}{A character vector indicating the names of the biomarkers included in the Phenotypic Age algorithm.}
}
\value{
An object of class "phenoage". This object is a list with two elements (data and fit). The dataset can be drawn by typing 'data'. The model can be drawn by typing 'fit'.
}
\description{
Train Phenotypic Age algorithm in NHANES III and project into NHANES IV.
}
\details{
Training Phenotypic Age algorithm using the NHANES III and projecting into NHANES IV dataset.
}
\examples{
#Phenoage using NHANES
phenoage = phenoage_nhanes(biomarkers=c("albumin_gL","lymph","mcv","glucose_mmol",
                           "rdw","creat_umol","lncrp","alp","wbc"))

#Extract phenoage dataset
data = phenoage$data


}
