% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ba.R
\name{plot_ba}
\alias{plot_ba}
\title{plot_ba}
\usage{
plot_ba(data, agevar, label)
}
\arguments{
\item{data}{The dataset for plotting correlations}

\item{agevar}{A character vector indicating the names of the biological aging measures}

\item{label}{A character vector indicating the labels of the biological age measures}
}
\description{
Plot correlations between biological aging measures and chronological age
}
\details{
Plot correlations between biological aging measures and chronological age
}
\note{
Chronological age and gender variables need to be named "age" and "gender"
}
\examples{
#Calculate phenoage
f1 = plot_ba(data = data, agevar = c("bioage", "phenoage", "hd"))

f1

}
