% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baa.R
\name{plot_baa}
\alias{plot_baa}
\title{plot_baa}
\usage{
plot_baa(data, agevar, label, axis_type)
}
\arguments{
\item{data}{The dataset for plotting corplot}

\item{agevar}{A character vector indicating the names of the interested biological aging measures}

\item{label}{A character vector indicating the labels of the biological aging measures
Values should be formatted for displaying along diagonal of the plot
Names should be used to match variables and order is preserved}

\item{axis_type}{A character vector indicating the axis type (int or float)
Use variable name to define the axis type}
}
\description{
Plot correlations between biological age advancement (BAA) and chronological age
}
\details{
Plot correlations between BAA and chronological age
}
\examples{
#Create corplot of BAA with chronologicl age
agevar = c("kdm_advance0",
           "phenoage_advance0",
           "kdm_advance",
           "phenoage_advance",
           "hd",
           "hd_log")

label = c("kdm_advance0"="KDM\nBiological\nAge",
           "phenoage_advance0"="Levine\nPhenotypic\nAge",
           "kdm_advance"="Modified-KDM\nBiological\nAge",
           "phenoage_advance"="Modified-Levine\nPhenotypic\nAge",
           "hd" = "Mahalanobis\nDistance",
           "hd_log" = "Log\nMahalanobis\nDistance")

axis_type = c("kdm_advance0"="float",
              "phenoage_advance0"="float",
              "kdm_advance"="float",
              "phenoage_advance"="flot",
              "hd"="flot",
              "hd_log"="float")

f2 = plot_baa(data, agevar, labels, axis_type)

f2

}
