% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_health.R
\name{table_health}
\alias{table_health}
\title{table_health}
\usage{
table_health(data, agevar, outcome, label)
}
\arguments{
\item{data}{The dataset for linear regression table}

\item{agevar}{A character vector indicating the names of the interested biological aging measures}

\item{outcome}{A character vector indicating the name of the interested current health status outcomes}

\item{label}{A character vector indicating the labels of the biological aging measures}
}
\description{
Association with current health status outcomes
}
\details{
Association with current health status outcomes, adjusting for chronological age and gender and stratified by gender, race, and age
}
\note{
Chronological age and gender variables need to be named "age" and "gender"
}
\examples{
table2 = table_health(data,
                      agevar = c("kdm_advance0","phenoage_advance0",
                               "kdm_advance","phenoage_advance",
                               "hd","hd_log"),
                      outcome = c("health","adl","lnwalk","grip_scaled"),
                      label = c("KDM\nBiological\nAge",
                                "Levine\nPhenotypic\nAge",
                                "Modified-KDM\nBiological\nAge",
                                "Modified-Levine\nPhenotypic\nAge",
                                "Mahalanobis\nDistance",
                                "Log\nMahalanobis\nDistance"))

table2$table
table2$n

}
