% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_health.R
\name{table_health}
\alias{table_health}
\title{table_health}
\usage{
table_health(data, agevar, outcome, label)
}
\arguments{
\item{data}{A dataset with projected biological aging measures for analysis.}

\item{agevar}{A character vector indicating the names of the biological aging measures.}

\item{outcome}{A character vector indicating the name of the healthspan-related characteristics.}

\item{label}{A character vector indicating the labels of the biological aging measures.}
}
\value{
The result is a list with two elements (table and  n). The regression table can be drawn by typing 'table'. The sample size table can be drawn by typing 'n'.
}
\description{
Associations of biological aging measures with healthspan-related characteristics.
}
\details{
Coefficients are from linear regressions of healthspan-related characteristics on biological aging measures. Outcome variables were standardized to have M=0, SD=1 for analysis. KDM Biological Age and Levine Phenotypic Age measures were differenced from chronological age for analysis (i.e. values = BA-CA). These differenced values were then standardized to have M=0, SD=1 separately for men and women within the analysis sample so that effect-sizes are denominated in terms of a sex-specific 1 SD unit increase in biological age advancement. Models included covariates for chronological age and sex.
}
\note{
Chronological age, gender, and race/ethnicity variables need to be named "age", "gender", and "race".
}
\examples{
table2 = table_health(data,
                      agevar = c("kdm_advance0","phenoage_advance0",
                               "kdm_advance","phenoage_advance",
                               "hd","hd_log"),
                      outcome = c("health","adl","lnwalk","grip_scaled"),
                      label = c("KDM\nBiological Age\nAdvancement",
                                "Levine\nPhenotypic Age\nAdvancement",
                                "Modified-KDM\nBiological Age\nAdvancement",
                                "Modified-Levine\nPhenotypic Age\nAdvancement",
                                "Homeostatic\nDysregulation",
                                "Log\nHomeostatic\nDysregulation"))

table2$table
table2$n

}
