% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_surv.R
\name{table_surv}
\alias{table_surv}
\title{table_surv}
\usage{
table_surv(data, agevar, time, status)
}
\arguments{
\item{data}{The dataset for plotting corplot}

\item{agevar}{A character vector indicating the names of the interested biological age measures}

\item{time}{A character vector (length=1) indicating the name of the variable for survival time}

\item{status}{A character vector (length=1) indicating the name of the variable for survival status}
}
\description{
Mortality validation for biological age measures
}
\details{
Mortality validation for biological age measures, adjusting for chronological age and gender and stratified by gender, race, and age
}
\note{
Chronological age, gender, and race/ethnicity variables need to be named "age", "gender", and "race"
}
\examples{
table1 = table_surv(nhanes,
                    agevar = c("bioage_advance0","phenoage_advance0",
                               "bioage_advance","phenoage_advance",
                               "hd","hd_log"),
                    time = "permth_exm",
                    status = "mortstat")

table1

}
